OTV: create forwarding package
This change creates a new package to house the forwarding service and
the vidforward service
