+++ cmd/oceantv/broadcast/config.go
@@ -1,3 +1,25 @@
+/*
+AUTHORS
+  Saxon Nelson-Milton <saxon@ausocean.org>
+
+LICENSE
+  Copyright (C) 2026 the Australian Ocean Lab (AusOcean)
+
+  This file is part of Ocean TV. Ocean TV is free software: you can
+  redistribute it and/or modify it under the terms of the GNU
+  General Public License as published by the Free Software
+  Foundation, either version 3 of the License, or (at your option)
+  any later version.
+
+  Ocean TV is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  GNU General Public License for more details.
+
+  You should have received a copy of the GNU General Public License
+  in gpl.txt. If not, see <http://www.gnu.org/licenses/>.
+*/
+
 package broadcast
 
 import (
@@ -13,7 +35,11 @@ import (
 	"github.com/ausocean/cloud/model"
 )
 
-const Scope = "Broadcast" // Scope under which broadcast configs are stored.
+const (
+	Scope                  = "Broadcast"                        // Scope under which broadcast configs are stored.
+	SecondaryPostfix       = "(Secondary)"                      // Post fix used on end of secondary broadcast names.
+	RTMPDestinationAddress = "rtmp://a.rtmp.youtube.com/live2/" // Base address for RTMP destination (RTMP key is appended).
+)
 
 // Config holds configuration data for a YouTube broadcast.
 type Config struct {
