+++ cmd/oceantv/broadcast_permanent.go
@@ -25,146 +25,3 @@ LICENSE
 */
 
 package main
-
-import (
-	"bytes"
-	"fmt"
-	"io"
-	"mime/multipart"
-	"net/http"
-	"strings"
-	"time"
-
-	"github.com/ausocean/cloud/model"
-)
-
-type SlateOption func(*Cfg) error
-
-type ForwardingService interface {
-	Stream(cfg *Cfg) error
-	Slate(cfg *Cfg, opts ...SlateOption) error
-	UploadSlate(cfg *Cfg, name string, file io.Reader) error
-}
-
-type vidforwardStatus string
-
-const (
-	vidforwardStatusPlay  vidforwardStatus = "play"
-	vidforwardStatusSlate vidforwardStatus = "slate"
-)
-
-type VidforwardService struct {
-	log func(string, ...interface{})
-}
-
-func NewVidforwardService(log func(string, ...interface{})) *VidforwardService {
-	return &VidforwardService{log}
-}
-
-func (v *VidforwardService) Stream(cfg *Cfg) error {
-	return vidforwardRequest(cfg, vidforwardStatusPlay, v.log)
-}
-
-type SlateType string
-
-const (
-	Default    SlateType = "default"
-	LowVoltage SlateType = "low-voltage"
-)
-
-// WithType is an option for the Slate function that allows the caller to specify
-// the type of slate to display.
-// This is currently just a stub.
-func WithType(slate SlateType) SlateOption {
-	return func(cfg *Cfg) error {
-		return nil
-	}
-}
-
-func (v *VidforwardService) Slate(cfg *Cfg, opts ...SlateOption) error {
-	return vidforwardRequest(cfg, vidforwardStatusSlate, v.log)
-}
-
-func (v *VidforwardService) UploadSlate(cfg *Cfg, name string, file io.Reader) error {
-	body := &bytes.Buffer{}
-
-	// Not closing this just yet, see close below.
-	writer := multipart.NewWriter(body)
-
-	formFileWriter, err := writer.CreateFormFile("slate-file", name)
-	if err != nil {
-		return fmt.Errorf("could not create form file writer: %w", err)
-	}
-
-	_, err = io.Copy(formFileWriter, file)
-	if err != nil {
-		return fmt.Errorf("could not copy slate file to formFileWriter: %w", err)
-	}
-
-	// We need to close the writer before we can make the request, otherwise
-	// we get a "multipart EOF" error on the other side.
-	err = writer.Close()
-	if err != nil {
-		return fmt.Errorf("could not close writer: %w", err)
-	}
-
-	req, err := http.NewRequest("POST", "http://"+cfg.VidforwardHost+"/slate", body)
-	if err != nil {
-		return fmt.Errorf("could not create new /slate request: %w", err)
-	}
-
-	req.Header.Set("Content-Type", writer.FormDataContentType())
-
-	const timeout = 10 * time.Second
-	client := &http.Client{Timeout: timeout}
-	resp, err := client.Do(req)
-	if err != nil {
-		return fmt.Errorf("could not do /slate request: %w, resp: %v", err, resp)
-	}
-	return nil
-}
-
-func vidforwardRequest(cfg *Cfg, status vidforwardStatus, log func(string, ...interface{})) error {
-	primary, secondary := cfg, cfg
-	var err error
-
-	// If the cfg is the secondary broadcast then we need to get the primary broadcast.
-	if strings.Contains(primary.Name, secondaryBroadcastPostfix) {
-		// We need to use broadcastByName to get the primary broadcast.
-		// This will mean that we'll need to trim off the secondaryBroadcastPostfix.
-		primaryName := strings.TrimSuffix(primary.Name, secondaryBroadcastPostfix)
-		primary, err = broadcastByName(primary.SKey, primaryName)
-		if err != nil {
-			return fmt.Errorf("could not get primary broadcast configuration: %w", err)
-		}
-
-		// Otherwise we just need to load the secondary broadcast.
-	} else {
-		secondary, err = broadcastByName(primary.SKey, primary.Name+secondaryBroadcastPostfix)
-		if err != nil {
-			return fmt.Errorf("could not get secondary broadcast configuration: %w", err)
-		}
-	}
-
-	urls := []string{rtmpDestinationAddress + primary.RTMPKey, rtmpDestinationAddress + secondary.RTMPKey}
-
-	data := struct {
-		MAC, Status string
-		URLs        []string
-	}{
-		MAC:    model.MacDecode(primary.CameraMac),
-		URLs:   urls,
-		Status: string(status),
-	}
-
-	log("attempting to update vidforward configuration, data: %+v", data)
-
-	// We're allowing some tolerance to failed requests here because it may be that we've
-	// caught vidforward during a service restart.
-	const maxRetries = 3
-	err = performRequestWithRetries("http://"+cfg.VidforwardHost+"/control", data, maxRetries, log)
-	if err != nil {
-		return fmt.Errorf("could not perform request with retries: %v", err)
-	}
-	return nil
-}
