+++ cmd/oceantv/broadcast_test.go
@@ -35,6 +35,7 @@ import (
 	"time"
 
 	"github.com/ausocean/cloud/cmd/oceantv/event"
+	"github.com/ausocean/cloud/cmd/oceantv/forwarding"
 	"github.com/ausocean/cloud/cmd/oceantv/ratelimit"
 	"github.com/ausocean/cloud/cmd/oceantv/yt"
 	"github.com/ausocean/cloud/datastore"
@@ -293,7 +294,7 @@ type dummyForwardingService struct{}
 
 func newDummyForwardingService() *dummyForwardingService                                  { return &dummyForwardingService{} }
 func (v *dummyForwardingService) Stream(cfg *Cfg) error                                   { return nil }
-func (v *dummyForwardingService) Slate(cfg *Cfg, opts ...SlateOption) error               { return nil }
+func (v *dummyForwardingService) Slate(cfg *Cfg, opts ...forwarding.SlateOption) error    { return nil }
 func (v *dummyForwardingService) UploadSlate(cfg *Cfg, name string, file io.Reader) error { return nil }
 
 type request struct {
