+++ cmd/oceantv/forwarding/service.go
@@ -1,0 +1,38 @@
+/*
+AUTHORS
+  Saxon Nelson-Milton <saxon@ausocean.org>
+
+LICENSE
+  Copyright (C) 2026 the Australian Ocean Lab (AusOcean)
+
+  This file is part of Ocean TV. Ocean TV is free software: you can
+  redistribute it and/or modify it under the terms of the GNU
+  General Public License as published by the Free Software
+  Foundation, either version 3 of the License, or (at your option)
+  any later version.
+
+  Ocean TV is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  GNU General Public License for more details.
+
+  You should have received a copy of the GNU General Public License
+  in gpl.txt. If not, see <http://www.gnu.org/licenses/>.
+*/
+
+package forwarding
+
+import (
+	"io"
+
+	"github.com/ausocean/cloud/cmd/oceantv/broadcast"
+)
+
+type SlateOption func(*broadcast.Config) error
+
+// Service is an interface to represent a service which handles forwarding.
+type Service interface {
+	Stream(cfg *broadcast.Config) error
+	Slate(cfg *broadcast.Config, opts ...SlateOption) error
+	UploadSlate(cfg *broadcast.Config, name string, file io.Reader) error
+}
