+++ cmd/oceantv/system.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/ausocean/cloud/cmd/oceantv/broadcast"
 	"github.com/ausocean/cloud/cmd/oceantv/event"
+	"github.com/ausocean/cloud/cmd/oceantv/forwarding"
 	"github.com/ausocean/cloud/cmd/oceantv/notifier"
 	"github.com/ausocean/cloud/cmd/oceantv/yt"
 	"github.com/ausocean/cloud/notify"
@@ -40,7 +41,7 @@ func withBroadcastService(bs Svc) broadcastSystemOption {
 	}
 }
 
-func withForwardingService(fs ForwardingService) broadcastSystemOption {
+func withForwardingService(fs forwarding.Service) broadcastSystemOption {
 	return func(bs *broadcastSystem) error {
 		bs.ctx.fwd = fs
 		return nil
@@ -140,7 +141,7 @@ func newBroadcastSystem(ctx Ctx, store Store, cfg *Cfg, logOutput func(v ...any)
 	bus := event.NewBasicEventBus(ctx, storeEventsAfterCtx, log)
 
 	// This context will be used by the state machines for access to our bits and bobs.
-	broadcastContext := &broadcastContext{cfg, man, store, svc, NewVidforwardService(log), bus, &revidCameraClient{}, logOutput, nil}
+	broadcastContext := &broadcastContext{cfg, man, store, svc, forwarding.NewVidforwardService(log, broadcastByName), bus, &revidCameraClient{}, logOutput, nil}
 
 	// Subscribe event handler that notifies on events that implement errorEvent.
 	bus.Subscribe(func(e event.Event) error {
